Factor out `CScriptStorage`, `CScriptEngine::process_file`
