+++ src/xrServerEntities/script_engine.h
@@ -8,11 +8,10 @@
 
 #pragma once
 
-#include "script_storage.h"
+#include "script_storage_space.h"
 #include "script_export_space.h"
 #include "script_space_forward.h"
 #include "associative_vector.h"
-#include "script_storage.h"
 
 //AVO: lua re-org
 #include "lua.hpp"
@@ -24,6 +23,32 @@
 //#define DBG_DISABLE_SCRIPTS
 
 #include "script_engine_space.h"
+
+#ifndef MASTER_GOLD
+#	define USE_DEBUGGER
+#	define USE_LUA_STUDIO
+#endif //-!MASTER_GOLD
+
+#ifdef XRGAME_EXPORTS
+#	ifndef MASTER_GOLD
+#		define PRINT_CALL_STACK
+#	endif //-!MASTER_GOLD
+#else //!XRGAME_EXPORTS
+#	ifndef NDEBUG
+#		define PRINT_CALL_STACK
+#	endif // #ifndef NDEBUG
+#endif //-XRGAME_EXPORTS
+
+//AVO: allow LUA debug prints (i.e.: ai().script_engine().script_log(ScriptStorage::eLuaMessageTypeError, "CWeapon : cannot access class member Weapon_IsScopeAttached!");)
+#include "..\build_config_defines.h"
+#ifndef DEBUG
+#   ifdef LUA_DEBUG_PRINT
+#       define PRINT_CALL_STACK
+#   endif
+#endif //-!DEBUG
+//-AVO
+
+using namespace ScriptStorage;
 
 class CScriptProcess;
 class CScriptThread;
@@ -44,10 +69,75 @@ struct lua_Debug;
 #	endif // #ifndef USE_LUA_STUDIO
 #endif
 
-class CScriptEngine : public CScriptStorage
+class CScriptEngine
 {
+private:
+	lua_State* m_virtual_machine;
+	CScriptThread* m_current_thread;
+	BOOL m_jit;
+
+#ifdef DEBUG
 public:
-	typedef CScriptStorage inherited;
+	bool						m_stack_is_ready;
+#endif //-DEBUG
+
+#ifdef LUA_DEBUG_PRINT//PRINT_CALL_STACK
+protected:
+	CMemoryWriter m_output;
+#else
+#   ifdef DEBUG
+protected:
+	CMemoryWriter m_output;
+#   endif //-DEBUG
+#endif //-LUA_DEBUG_PRINT PRINT_CALL_STACK
+
+protected:
+	static int vscript_log(ScriptStorage::ELuaMessageType tLuaMessageType, LPCSTR caFormat, va_list marker);
+	bool do_file(LPCSTR caScriptName, LPCSTR caNameSpaceName);
+	void reinit();
+
+public:
+	//#ifdef PRINT_CALL_STACK
+	void print_stack();
+	//AVO: added to stop duplicate stack output prints in log
+	static int __cdecl script_log_no_stack(ScriptStorage::ELuaMessageType tLuaMessageType, LPCSTR caFormat, ...);
+	//-AVO
+	//#endif //-PRINT_CALL_STACK
+
+public:
+	CScriptEngine();
+	~CScriptEngine();
+	IC lua_State* lua();
+	IC void current_thread(CScriptThread* thread);
+	IC CScriptThread* current_thread() const;
+	int compile_buffer(
+		lua_State* L,
+		std::string caString,
+		LPCSTR caScriptName,
+		LPCSTR caNameSpaceName = 0
+	);
+	int load_buffer(
+		lua_State* L,
+		LPCSTR caBuffer,
+		size_t tSize,
+		LPCSTR caScriptName,
+		LPCSTR caNameSpaceName = 0
+	);
+	bool load_file_into_namespace(LPCSTR caScriptName, LPCSTR caNamespaceName);
+	bool namespace_loaded(LPCSTR caName, bool remove_from_stack = true);
+	luabind::object name_space(LPCSTR namespace_name);
+	int error_log(LPCSTR caFormat, ...);
+	static int __cdecl script_log(ELuaMessageType message, LPCSTR caFormat, ...);
+	static bool print_output(lua_State* L, LPCSTR caScriptName, int iErorCode = 0);
+	static void print_error(lua_State* L, int iErrorCode);
+	void on_error(lua_State* L);
+
+#ifdef LUA_DEBUG_PRINT //DEBUG
+public:
+	void flush_log();
+#endif //-LUA_DEBUG_PRINT DEBUG
+
+public:
 	typedef ScriptEngine::EScriptProcessors EScriptProcessors;
 	typedef associative_vector<EScriptProcessors, CScriptProcess*> CScriptProcessStorage;
 
@@ -74,10 +164,8 @@ private:
 	void add_no_file(LPCSTR file_name, u32 string_length);
 
 public:
-	CScriptEngine();
-	virtual ~CScriptEngine();
 	void init();
-	virtual void unload();
+	void unload();
 	static int lua_panic(lua_State* L);
 	static void lua_error(lua_State* L);
 	static int lua_pcall_failed(lua_State* L);
@@ -86,14 +174,12 @@ public:
 #endif // #ifdef DEBUG
 	void setup_callbacks();
 	void load_common_scripts();
-	bool load_file(LPCSTR caScriptName, LPCSTR namespace_name);
 	IC CScriptProcess* script_process(const EScriptProcessors& process_id) const;
 	IC void add_script_process(const EScriptProcessors& process_id, CScriptProcess* script_process);
 	void remove_script_process(const EScriptProcessors& process_id);
 	void setup_auto_load();
+	bool load_package(LPCSTR file_name, bool warn_if_not_exist = true);
 	void unload_package(LPCSTR package);
-	bool process_file_if_exists(LPCSTR file_name, bool warn_if_not_exist);
-	void process_file(LPCSTR file_name);
 protected:
 	bool object(LPCSTR caIdentifier, int type);
 	bool object(LPCSTR caNamespaceName, LPCSTR caIdentifier, int type);
@@ -117,7 +203,6 @@ public:
 	inline cs::lua_studio::world* debugger					() const { return m_lua_studio_world; }
 #	endif // ifndef USE_LUA_STUDIO
 #endif
-	virtual void on_error(lua_State* state);
 	void collect_all_garbage();
 
 DECLARE_SCRIPT_REGISTER_FUNCTION
