+++ src/xrServerEntities/script_engine_inline.h
@@ -8,6 +8,22 @@
 
 #pragma once
 
+IC lua_State* CScriptEngine::lua()
+{
+	return (m_virtual_machine);
+}
+
+IC void CScriptEngine::current_thread(CScriptThread* thread)
+{
+	VERIFY((thread && !m_current_thread) || !thread);
+	m_current_thread = thread;
+}
+
+IC CScriptThread* CScriptEngine::current_thread() const
+{
+	return (m_current_thread);
+}
+
 IC void CScriptEngine::add_script_process(const EScriptProcessors& process_id, CScriptProcess* script_process)
 {
 	//	CScriptProcessStorage::const_iterator	I = m_script_processes.find(process_id);
